Tests & Routes - Refactor: Removed mockIf to mockOnce. Moved all route file with it tests to a folder with it's name
