+++ projects/Stronghold/src/routes/bookmanager/index.ts
@@ -1,0 +1,1 @@
+export * from './bookmanager';
