+++ projects/Stronghold/src/routes/token/index.ts
@@ -1,0 +1,1 @@
+export * from './token';
