+++ projects/Stronghold/src/routes/token/token.spec.ts
@@ -1,9 +1,9 @@
 import request from 'supertest';
 import fetchMock from 'jest-fetch-mock';
 
-import { databaseTestModule } from "../tests/database";
-import { testsLogger } from "../tests/testsLogger";
-import app from "../app";
+import { databaseTestModule } from "../../tests/database";
+import { testsLogger } from "../../tests/testsLogger";
+import app from "../../app";
 
 const log = testsLogger.extend('routes:token');
 
@@ -26,21 +26,9 @@ describe('Person Mutations', () => {
     };
 
     beforeEach(() => {
+
         // if you have an existing `beforeEach` just add the following lines to it
-        fetchMock.mockIf(/^https?:\/\/localhost:3233.*$/, async req => {
-
-            log('entered fetchMock mockIf fn');
-            if (req.url.endsWith('/token')) {
-                log('request to personssector');
-                return JSON.stringify({test: 'body'})
-            } else {
-
-                return {
-                    status: 404,
-                    body: JSON.stringify({error: 'Not Found'})
-                }
-            }
-        });
+        fetchMock.mockOnce(JSON.stringify({teste: 'teste'}));
     });
     beforeAll(() => connect());
     afterEach(async () => {
@@ -58,6 +46,22 @@ describe('Person Mutations', () => {
             {
                 authorization: 'Bearer 11111'
             }
+        );
+
+        log('response.body: ', response.body);
+        log('response.status: ', response.status);
+        log('response.header: ', response.header);
+        expect(response.body.test).toBe('body');
+        expect(response.status).toBe(200);
+    });
+
+    it('should return error object password grant', async () => {
+
+        log('token.spec');
+
+        const response = await requestFn(
+            {grant_type: 'password', username: 'test', password: '123'},
+            {}
         );
 
         log('response.body: ', response.body);
