+++ projects/Stronghold/src/routes/visitor/index.ts
@@ -1,0 +1,1 @@
+export * from './visitor';
