+++ projects/Stronghold/src/routes/visitor/visitor.spec.ts
@@ -1,9 +1,9 @@
 import request from 'supertest';
 import fetchMock from 'jest-fetch-mock';
 
-import { databaseTestModule } from "../tests/database";
-import { testsLogger } from "../tests/testsLogger";
-import app from "../app";
+import { databaseTestModule } from "../../tests/database";
+import { testsLogger } from "../../tests/testsLogger";
+import app from "../../app";
 
 const log = testsLogger.extend('routes:visitor');
 
@@ -27,21 +27,9 @@ describe('Visitor Route', () => {
     };
 
     beforeEach(() => {
-        // if you have an existing `beforeEach` just add the following lines to it
-        fetchMock.mockIf(/^https?:\/\/localhost:3233.*$/, async req => {
 
-            log('entered fetchMock mockIf fn');
-            if (req.url.endsWith('/visitor')) {
-                log('request to personssector');
-                return JSON.stringify({test: 'body'})
-            } else {
-
-                return {
-                    status: 404,
-                    body: JSON.stringify({error: 'Not Found'})
-                }
-            }
-        });
+        fetchMock.mockOnce(JSON.stringify({teste: 'teste'}));
+        fetchMock.mockOnce(JSON.stringify({teste: 'testeeeeeeee'}));
     });
     beforeAll(() => connect());
     afterEach(async () => {
